requirements.txt
click==7.1.2
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.11
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-decouple==3.3
smmap==3.0.4
Werkzeug==1.0.1
